Workaround: temporary replacement of elements in detector materials.
Apparently GENIE, after its coming of age to version 3, has become
strict about the elements in the geometry description that are not
associated with any cross section.
I have created a "light" geometry replacing two such elements,
and set the new geometry as default.
GEANT4 is not affected by the change.