Updated overburden configurations.
Configurations are automatically generated via
scripts/generateOverburdenConfigurations.sh
which creates two configurations (with and without overburden)
for each of the configuration file names listed in fcl/OverburdenConfigurations.txt.

Remember that after new configurations are added, they need to be added to GIT reposiory
and if they lie in a new subdirectory, that subdirectory must contain a new
`CMakeLists.txt` (`install_fhicl()`) and be linked to the parent directory
(`add_subdirectory("newSubDir")` in the parent directory `CMakeLists.txt`).