Revised random number policy for ICARUS.
Since `perEvent` is not in general safe, `random` was set as new default.
This commit sets the policy to `perEvent` for the standard simulation
service set (which includes G4 and DetSim). In addition, generators
using only RadioGen or SingleGen are set back to `perEvent`,
and G4 processes which include generation (e.g. "intime") are set to
`random`. Reco configuration happens to happily inherit the simulation
configuration (but Reco does not use random numbers anyway, does it??).