Trigger efficiency algorithm updates.
This is a horrible monolitic commit that includes the work of several months by three people
(Jacob Zettlemoyer, Animesh Chatterjee and Gianluca Petrillo)
and covers many aspects of trigger: emulation algorithms,
extension to emulate on arbitrary trigger gates and around track times,
configurations to run emulation in data and simulation,
and various utilities supporting any of these.

These changes interleave with other developments in PMT simulation and reconstruction,
too many and too carelessly committed for a commit-by-commit pick.
So here is the monolitic commit that looks like a single author got bit by a spider
and started monkeying code for three weeks.
With the apologies of that same "author".