Lowered our standards to make Clang happy
Clang uses unsigned integrals for unscoped enumerators (like LArSoft's `sim::BeamType_t`)
and the prescribed `int` (signed) for the scoped ones.
This implies narrowing when trying to cast a value of the former type into the latter.
GCC uses signed types for unscoped enumerators so it does not present the issue.