Update CRTHitRecoAlg.cc
Added an if statement to fix how the CRT Hit timestamps are averaged when there are >10 CRT DataProducts entering into the CRT Hit. Adding 11 or more 18 digit unix timestamps to be averaged in lines 1136-1137 can overload the maximum value of a uint64_t, giving a non-sensible timestamp for T0. As a work around we add and average just the nanosecond portion of the unix timestamp and then add onto the second portion to get a full averaged unix timestamp.