Comment out side CRT GT calculation addition and continue to use Top CRT GT as reference for side CRT Hits
For this PR, im now removing the Side CRT Global trigger calculation. Although adding this was the original motivation of this PR, I've discovered some offsets I need to understand and will not be ready for the production release. I would still like this PR to be merged because i've addressed many comments by @PetrilloAtWork related to other details of the CRTHitRecoAlg.cc. I will make an additional PR with this two suggestions:
* move trigger offset to be a configurable parameter
* change formatting of T0 and T1 delay structure in fcl, instead of having two structures with 2 items having one structure with both T0 and T1 delays saved as suggested by F. Poppi
as well as hopefully a fix to the Side CRT GT calculation, but we can use the Top CRT GT calculation for the production release.