Update crthitproducer.fcl
remove my addition of the T0 and T1 line delays here as they were giving CI failures in the stage0 configuration. i believe this was due to no FEB_T*delay_side tables saved in the standard crtsimhitproducer configuration, and the CRTHitReco algorithm calls for these tables. I think in the future for a work around we will want a fcl parameter for whether to expect MC or data, and to only load in the cable delays when running the algorithm on data. i can open an issue for this.