GnocchiCalorimetry optimization: replaced linear search with logarithmic
The search of the bin from a point has been changed from a linear time to logarithmic.
The bins have been sorted with an arbitrary criterion.
Additional comparison operations compatible with that criterion are defined to compare a bin and a point.
If a point is contained in a bin earlier than the comparison one, the poiny is labelled smaller than the comparison bin.
If a point is contained in a bin later than the comparison one, the poiny is labelled greater than the comparison bin.
If a point is contained in the comparison bin, it's labelled neither smaller nor greater, thus equal, to the bin.
Leveraging on this infrastructure, the standard binary search is used to look up the bin.