Fix a bug in PMT digitization
The algorithm creating readout buffers (future `raw::OpDetWaveform`) from the continuous readout was written expecting (and asserting) that a trigger primitive can't be at the first tick of the enable gate.
This was a reasonable assumption when the trigger algorithm looked for a transition from the previous tick from below to above threshold.
Now an alternative algorithm is also available which considers just the value of the sample compared to the threshold, so that expectation is violated and the assertion failed.
A small change in the algorithm makes it robust to that rare occurrence.