CRTPMTMatchingProducer: trigger-related code more robust
Trigger information is used only to determine the in-beam and in-gate flag values
for storage in the data product: a missing trigger information is acceptable
but it will set that information to `false`.
The code now makes sure that this happen and prints a warning when it does.
Note that if a Monte Carlo sample has not run the trigger simulation,
the outcome will be equivalent to not having trigger information.
The code is now using the specific trigger utilities to manage the gate types
(sbn::triggerSource), which makes it more readable.

Nonetheless, the resulting data product is effectively the same as before.