CRTPMTMatching: turned all times on data members and most interfaces into microseconds
Some data products (e.g. optical reconstruction ones: hits and flashes)
live in LArSoft and follow LArSoft conventions, which demand times in microseconds.
Some other information (like timestamps) is naturally coming in nanoseconds.
The CRT/PMT matching mixes the two, imposing some choice.
The choice in this commit is to move everything to LArSoft convention (microseconds),
accepting the weirdness of expressing things like times of flight in that unit
even if it's not at all natural for them. The gain is predictable data and interface.