PMT and trigger support for Stage0/Stage1 data
It was decided that Stage0 data would only serialize the PMT waveforms around the beam gate,
and these waveforms are a new data product from the point of view of the framework.
To make most of the algorithm still work, additional information must be saved.
That includes making proper associations of the existing data with the new waveforms
(like baselines) and/or recreate it for them (e.g. metadata).
This commit introduces a module to perform a reassociation of the new waveforms
to the existing baseline data product, in case it was not there yet (it actually was,
but bugged; the bug is fixed with a previous commit, but the existing data suffers from it),
a new filter module to verify that the waveforms are actually there (workaround for
incomplete events) and adds resiliency to one of the trigger algorithms in case
of missing data as well.