Updated support for trigger emulation on data
Syntax of LVDSgates threshold configuration has been uniformed to the other
module (even if the interpretation under the hood has still to be different).
TriggerSimulationOnGates can specify which time reference the gates use.
This is necessary because it turns out that the trigger decoder is erroneously
using the trigger time instead of the standard beam gate time
(this will be the subject of a different fix; the change is necessary to
correctly emulate triggers in the already decoded data).