Workaround on FHiCL parameter set bug
The bug is discussed in issue #14 of art-framework-suite/fhicl-cpp
(https://github.com/art-framework-suite/fhicl-cpp/issues/14).
The workaround is that `util::readConfigurationFromArtPrincipal`
t emits a warning if an expected configuration is not present,
instead of throwing an exception. That puts the load on the caller
for checking that the configuration that is needed is actually
present (in some of the cases we have seen, it was, so no harm done).