Adjusting the non-refactored workflow
The second pass of the trigger simulation has been moved back to stage0 to remain consistent with standard workflows.
The stage0 fhicl has been updated to point the various producers, which previously expected opdaq to the new shiftIonization label.
Finally, it was decided that it would be useful to retain the old raw::Trigger product, so this is no longer dropped and is labeled under "emuTriggerUnshifted".