Trigger decoder (V3) now uses channel mapping database for decoding bits
PMT pair bits are assigned to their proper position via a mapping informed
by the PMT channel mapping database (which requires an update).
In case the information is not available (old database), an option can be
activated to allow for a legacy hard-coded fallback. Note that this legacy
mapping is known to be incorrect for Run2.
The use of the database is enabled by default.
A utility to map LVDS bits and PMT channels is also provided (and used
by the decoder).
Adder bits are not yet mapped in this way.