Trigger decoder: fixed condition for trigger from both cryostats
It has been found that the FPGS logic assigning a trigger to adder or
majority logic may, in case the decision is "both", miss one bit,
hence returning a value 3 or 5 instead of the canonical 7.
Although this has not been observed in the triggering cryostat information,
I am relaxing the decoding logic so that in those cases the trigger
is assigned to both cryostats.