TriggerSimulationOnGates: added transition-based trigger emulation
A variant of the trigger logic algorithm has been implemented,
that has been in production in Run3 since run 11821: a trigger can be emitted
only when there is a transition from requirements not met to met,
and the result of the transition is in the "beam" gate.

In addition, the emulation code has been moved by time-based to tick-based,
which is more natural, immediate and easy (the original choice was plain bad,
and I basically knew since the beginning).

Workarounds for PR LArSoft/lardataalg#44 have also been removed
(it would deserve its own commit, but it's too late to disentangle).